feat: refactor API URL retrieval and improve token decoding with type safety
