+++ apps/web/src/lib/api-client.ts
@@ -57,15 +57,6 @@ interface RequestOptions<TRequestBody> {
 }
 
 /**
- * Returns the base URL for the API
- */
-export function getApiUrl(): string {
-  return process.env.NEXT_PUBLIC_API_URL
-    ? String(`${process.env.NEXT_PUBLIC_API_URL}/api`)
-    : "http://localhost:3333/api";
-}
-
-/**
  * Custom error class for API-related errors
  */
 export class ApiError extends Error {
