+++ apps/web/src/lib/api-url.ts
@@ -1,0 +1,8 @@
+/**
+ * Returns the base URL for the API
+ */
+export function getApiUrl(): string {
+  return process.env.NEXT_PUBLIC_API_URL
+    ? String(`${process.env.NEXT_PUBLIC_API_URL}/api`)
+    : "http://localhost:3333/api";
+}
